feat: phone login, thermal printing, branding, and UI polish
Add Firebase phone auth with Zoho session bind, ESC/POS Bluetooth
thermal printing, app icons/logo, settings/profile screens, multi-UOM
line editor improvements, document numbering, shipment orders report,
and dashboard list/grid layout with sales trend chart.

+++ lib/ui/core/cubit/line_editor_state.dart
@@ -1,8 +1,11 @@
 import 'package:equatable/equatable.dart';
 
 /// Reactive line-item calculation state for [LineEditorCubit].
+///
+/// [quantity] and [rate] are expressed in the currently selected unit;
+/// the math is unit-agnostic (subtotal = rate × quantity either way).
 class LineEditorState extends Equatable {
-  final int quantity;
+  final double quantity;
   final double rate;
   final double discount;
   final double taxPercentage;
@@ -24,7 +27,7 @@ class LineEditorState extends Equatable {
   });
 
   LineEditorState copyWith({
-    int? quantity,
+    double? quantity,
     double? rate,
     double? discount,
     double? taxPercentage,
